+++ src/Liquid.Dataverse/Liquid.Dataverse.csproj
@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 	  <Authors>Avanade Brazil</Authors>
 	  <Company>Avanade Inc.</Company>
 	  <Product>Liquid - Modern Application Framework</Product>
 	  <Copyright>Avanade 2019</Copyright>
-	  <Version>6.0.0</Version>
+	  <Version>8.0.0-rc-01</Version>
 	  <GenerateDocumentationFile>true</GenerateDocumentationFile>
 	  <IsPackable>true</IsPackable>
 	  <Description>Adapter for Microsoft Dataverse integrations.
@@ -18,9 +18,10 @@ This component is part of Liquid Application Framework.</Description>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.61.3" />
-    <PackageReference Include="Microsoft.PowerPlatform.Dataverse.Client" Version="1.1.22" />
-    <PackageReference Include="System.Security.Cryptography.Pkcs" Version="6.0.4" />
+    <PackageReference Include="Liquid.Core" Version="8.0.0-rc-01" />
+    <PackageReference Include="Microsoft.Identity.Client" Version="4.71.1" />
+    <PackageReference Include="Microsoft.PowerPlatform.Dataverse.Client" Version="1.2.5" />
+    <PackageReference Include="System.Security.Cryptography.Pkcs" Version="8.0.1" />
     <PackageReference Include="System.Text.RegularExpressions" Version="4.3.1" />
   </ItemGroup>
