+++ src/Liquid.Dataverse/logo.png
