+++ test/Liquid.Dataverse.Tests/LiquidDataverseTests.cs
@@ -6,38 +6,33 @@ using Microsoft.Xrm.Sdk.Metadata;
 using Microsoft.Xrm.Sdk.Query;
 using NSubstitute;
 using NSubstitute.ReceivedExtensions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Liquid.Adapter.Dataverse.Tests
+namespace Liquid.Dataverse.Tests
 {
-    public class DataverseAdapterTests
+    public class LiquidDataverseTests
     {
         private readonly IOrganizationServiceAsync _client;
-        private readonly ILiquidDataverseAdapter _sut;
-        public DataverseAdapterTests()
+        private readonly ILiquidDataverse _sut;
+        public LiquidDataverseTests()
         {
             var clientFactory = Substitute.For<IDataverseClientFactory>();
             _client = Substitute.For<IOrganizationServiceAsync>();
 
             clientFactory.GetClient().Returns(_client);
 
-            _sut = new DataverseAdapter(clientFactory);
+            _sut = new LiquidDataverse(clientFactory);
         }
 
         [Fact]
         public void Ctor_WhenClientFactoryIsNull_ThrowArgumentNullException()
         {
-            Assert.Throws<ArgumentNullException>(() => new DataverseAdapter(null));
+            Assert.Throws<ArgumentNullException>(() => new LiquidDataverse(null));
         }
 
         [Fact]
         public async Task GetById_WhenClientReturnResults_ReturnEntity()
         {
-            _client.RetrieveAsync(Arg.Any<string>(), Arg.Any<Guid>(),Arg.Any<ColumnSet>()).Returns(new Entity());
+            _client.RetrieveAsync(Arg.Any<string>(), Arg.Any<Guid>(), Arg.Any<ColumnSet>()).Returns(new Entity());
 
             var guidId = Guid.NewGuid();
 
@@ -89,7 +84,7 @@ namespace Liquid.Adapter.Dataverse.Tests
             _client.ExecuteAsync(Arg.Any<SetStateRequest>()).Returns(new OrganizationResponse());
 
             var entity = new EntityReference();
-            
+
             await _sut.SetState(entity, "1234", "1212");
 
             await _client.Received(1).ExecuteAsync(Arg.Any<SetStateRequest>());
@@ -107,35 +102,35 @@ namespace Liquid.Adapter.Dataverse.Tests
             await _client.Received(1).ExecuteAsync(Arg.Any<UpsertRequest>());
         }
 
-		[Fact]
-		public async Task Update_WithOptions_UseRightOrganizationRequestType_RequestParameters()
-		{
-			_client.UpdateAsync(Arg.Any<Entity>()).Returns(Task.CompletedTask);
+        [Fact]
+        public async Task Update_WithOptions_UseRightOrganizationRequestType_RequestParameters()
+        {
+            _client.UpdateAsync(Arg.Any<Entity>()).Returns(Task.CompletedTask);
 
-			var updatedEntity = new Entity();
+            var updatedEntity = new Entity();
 
-			await _sut.Update(updatedEntity, true, true, true);
+            await _sut.Update(updatedEntity, true, true, true);
 
-			await _client.Received(1).ExecuteAsync(Arg.Is<UpdateRequest>(ur => 
+            await _client.Received(1).ExecuteAsync(Arg.Is<UpdateRequest>(ur =>
             ur.Parameters.ContainsKey("BypassCustomPluginExecution")
             && ur.Parameters.ContainsKey("SuppressCallbackRegistrationExpanderJob")
-			&& ur.Parameters.ContainsKey("SuppressDuplicateDetection")
-			));
-		}
+            && ur.Parameters.ContainsKey("SuppressDuplicateDetection")
+            ));
+        }
 
-		[Fact]
-		public async Task DeleteById_WithOptions_UseRightOrganizationRequestType_RequestParameters()
-		{
+        [Fact]
+        public async Task DeleteById_WithOptions_UseRightOrganizationRequestType_RequestParameters()
+        {
             _client.DeleteAsync(Arg.Any<string>(), Arg.Any<Guid>()).Returns(Task.CompletedTask);
 
-			var guidId = Guid.NewGuid();
+            var guidId = Guid.NewGuid();
 
-			await _sut.Delete(guidId, "entityname", true, true);
+            await _sut.Delete(guidId, "entityname", true, true);
 
-			await _client.Received(1).ExecuteAsync(Arg.Is<DeleteRequest>(dr => 
+            await _client.Received(1).ExecuteAsync(Arg.Is<DeleteRequest>(dr =>
             dr.Parameters.ContainsKey("BypassCustomPluginExecution")
             ));
-		}
+        }
 
         [Fact]
         public async Task Create_WithOptions_UseRightOrganizationRequestType_RequestParameters()
@@ -144,8 +139,8 @@ namespace Liquid.Adapter.Dataverse.Tests
 
             var result = await _sut.Create(new Entity(), false, false, true);
 
-            _client.Received(1).Execute(Arg.Is<CreateRequest>(cr =>   
-            cr.Parameters.ContainsKey("BypassCustomPluginExecution") 
+            _client.Received(1).Execute(Arg.Is<CreateRequest>(cr =>
+            cr.Parameters.ContainsKey("BypassCustomPluginExecution")
             && cr.Parameters.ContainsKey("SuppressCallbackRegistrationExpanderJob")
             && cr.Parameters.ContainsKey("SuppressDuplicateDetection")
             ));
